Added test and code for requests that do not have a transfer-encoding.
If a baked request does not have a transfer-encoding then 
after the request is read and parsed it shouldn't have one.
The library previously set the transfer-encoding to nil
prior to reading the request, this results in failures with
Mechanize due to a nil header being left in the headers.