Raise a more appropriate exception when given an unknown response object
The change in responder.rb should be backwards-compatible (although it seems
unlikely that anyone's depending on it) because ArgumentError < StandardError.
The change in net_http.rb won't affect anyone because it's an internal sanity
check -- the guard in responder.rb catches everything first.

Also, a little related cleanup and test coverage.