Raise if .register_uri isn't given an options hash
The existing test below, #register_uri_without_domain_name, was
mistakenly constructed to call FakeWeb.register_uri with a string for
the third argument rather than an options hash as in all other
invocations. This was masked by the invalid second argument (trying to
use 'test_example2.txt' as a domain name) because FakeWeb checks that
first.

As of Ruby 2.2, URI supports parsing domain names that use the new
generic top-level domains (gTLDs), such as .info, .tech, etc.  Although
.txt is not an active TLD, URI no longer concerns itself with this
problem and now only raises a URI::InvalidURIError when a domain name is
syntactically invalid. So starting with Ruby 2.2, the error in this test
was no longer masked.

It started raising this instead:

    TypeError(<no implicit conversion of Symbol into Integer>)

...because it was trying to treat the String argument as a Hash
internally.

Since that error isn't very intuitive, this adds an explicit check for
passing non-hashy objects where a hash of response options is expected,
raising an ArgumentError instead.

Thanks for reporting: @strzibny in #53, @davidcornu in #58.