Tests: fix URI to really be invalid for Ruby 2.2
As of Ruby 2.2, URI supports parsing domain names that use the new
generic top-level domains (gTLDs), such as .info, .tech, etc.  Although
.txt is not an active TLD, URI no longer concerns itself with this
problem and now only raises a URI::InvalidURIError when a domain name is
syntactically invalid. So this test only worked on older Rubies.

Thanks to @davidcornu for reporting and suggesting a new string to use
in #58. Thanks also to @strzibny for reporting in #53.