Update Net::BufferedIO patch for Ruby 2.4-compat
In Ruby 2.4, Net::BufferedIO's method signature changed to:

    def initialize(io, read_timeout: 60, continue_timeout: nil, debug_output: nil)

Before that, it was just:

    def initialize(io)

Our around-advice-style monkey-patch wraps this method, so it needs to pass the
arguments through transparently. Ideally, we could just do this:

    def initialize_with_fakeweb(*args, **opts)

But this causes a syntax error on earlier Rubies. Using an eval like this is the
only way I can think of to work around this.

This fixes exceptions like:

    NoMethodError: undefined method `<<' for {:read_timeout=>60,
      :continue_timeout=>nil, :debug_output=>nil}:Hash

Thanks to @SamMolokanov for reporting in #60.